fix(file_name): remove useless solver
