--- Sudoku.Solver_TS_RM/Resources/utils.py
@@ -1,32 +1,0 @@
-
-instance = ((0,0,0,0,9,4,0,3,0),
-            (0,0,0,5,1,0,0,0,7),
-            (0,8,9,0,0,0,0,4,0),
-            (0,0,0,0,0,0,2,0,8),
-            (0,6,0,2,0,1,0,5,0),
-            (1,0,2,0,0,0,0,0,0),
-            (0,7,0,0,0,0,5,2,0),
-            (9,0,0,0,6,5,0,0,0),
-            (0,4,0,9,7,0,0,0,0))
-
-
-r = ((0,0,0,0,0,0,0,0,4),
-     (0,0,0,5,1,0,0,0,7),
-     (0,8,9,0,0,0,0,4,0),
-     (0,0,0,0,0,0,2,0,8),
-     (0,6,0,2,0,1,0,5,0),
-     (1,0,2,0,0,0,0,0,0),
-     (0,7,0,0,0,0,5,2,0),
-     (9,0,0,0,6,5,0,0,0),
-     (0,4,0,9,7,0,0,0,0))
-def troc():
-    s =    ((0,0,0,0,0,0,0,0,0),
-           (0,0,0,5,1,0,0,0,7),
-           (0,8,9,0,0,0,0,4,0),
-           (0,0,0,0,0,0,2,0,8),
-           (0,6,0,2,0,1,0,5,0),
-           (1,0,2,0,0,0,0,0,0),
-           (0,7,0,0,0,0,5,2,0),
-           (9,0,0,0,6,5,0,0,0),
-           (0,4,0,9,7,0,0,0,0))
-    return s
