--- Sudoku.Solver_TS_RM/Sudoku.DemoSolver.csproj
@@ -1,46 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="pythonnet" Version="3.1.0-preview2023-02-03" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Sudoku.Shared\Sudoku.Shared.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Update="Resource1.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>Resource1.resx</DependentUpon>
-    </Compile>
-    <Compile Remove="Resources_python.Designer.cs" />
-    <Compile Remove="Resources_python\**" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Update="Resource1.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Resource1.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
-    <EmbeddedResource Remove="Resources_python.resx" />
-    <EmbeddedResource Remove="Resources_python\**" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Remove="Resources\search.py" />
-    <None Remove="Resources\EmptyPythonSolver.py" />
-    <None Remove="Resources_python\**" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Include="Resources\EmptyPythonSolver.py" />
-  </ItemGroup>
-
-</Project>
